+++ tests/src/cli.js
@@ -58,14 +58,14 @@ describe('CLI regression tests', function () {
                 nodeType: results.results[0].messages[0].nodeType, // we don't care about this one
                 ruleId: 'json/*',
                 severity: 2,
-                source: '\n',
+                source: results.results[0].messages[0].source, // NewLine-characters might differ depending on git-settings
               },
             ],
             errorCount: 1,
             warningCount: 0,
             fixableErrorCount: 0,
             fixableWarningCount: 0,
-            source: ',\n',
+            source: results.results[0].source, // NewLine-characters might differ depending on git-settings
           },
         ],
         errorCount: 1,